+++ Core/Render/OpenGL/Renderers/Legacy/World/Entities/EntityProgram.cs
@@ -46,6 +46,8 @@ public class EntityProgram : RenderProgram
     private readonly int m_wallClipTextureLocation;
     private readonly int m_lineHeightsTextureLocation;
     private readonly int m_useBrightmapsLocation;
+    private readonly int m_downScaleAmountLocation;
+    private readonly int m_downScaleSampleFactorLocation;
 
     public EntityProgram(string name) : base($"Entity - {name}")
     {
@@ -86,6 +88,8 @@ public class EntityProgram : RenderProgram
         m_wallClipTextureLocation = Uniforms.GetLocation("wallClipTexture");
         m_lineHeightsTextureLocation = Uniforms.GetLocation("lineHeightsTexture");
         m_useBrightmapsLocation = Uniforms.GetLocation("useBrightmaps");
+        m_downScaleAmountLocation = Uniforms.GetLocation("downScaleAmount");
+        m_downScaleSampleFactorLocation = Uniforms.GetLocation("downScaleSampleFactor");
     }
     
     public void BoundTexture(TextureUnit unit) => ProgramUniforms.Set(unit, m_boundTextureLocation);
@@ -126,6 +130,8 @@ public class EntityProgram : RenderProgram
     public void CheckPlaneClip(bool value) => ProgramUniforms.Set(value, m_checkPlaneClipLocation);
     public void HealthBarMode(bool value) => ProgramUniforms.Set(value, m_healthBarModeLocation);
     public void UseBrightmaps(bool value) => ProgramUniforms.Set(value, m_useBrightmapsLocation);
+    public void SetSpriteClipDownScaleAmount(int value) => ProgramUniforms.Set(value, m_downScaleAmountLocation);
+    public void SetSpriteClipDownScaleSampleFactor(Vec2F value) => ProgramUniforms.Set(value, m_downScaleSampleFactorLocation);
 
     private const string BoxDefines = @"
         const float BoxWidth = 20;
@@ -353,6 +359,8 @@ public class EntityProgram : RenderProgram
         uniform int useBrightmaps;
         uniform vec3 viewPos;
         uniform float timeFrac;
+        uniform int downScaleAmount;
+        uniform vec2 downScaleSampleFactor;
 
         uniform sampler2D planeClipTexture;
         uniform sampler2D wallClipTexture;
@@ -381,7 +389,8 @@ public class EntityProgram : RenderProgram
         }
 
         bool discardPlaneClip() {
-            ivec2 getCoords = ivec2(gl_FragCoord.xy);
+            vec2 maxBounds = vec2(screenBounds.x / downScaleAmount - 1, screenBounds.y / downScaleAmount - 1);
+            ivec2 getCoords = ivec2(clamp(gl_FragCoord.xy / downScaleSampleFactor, vec2(0.0), maxBounds));
             vec3 wallClip = texelFetch(wallClipTexture, getCoords, 0).rgb;
             vec3 planeClip = texelFetch(planeClipTexture, getCoords, 0).rgb;