+++ Core/Render/OpenGL/Renderers/Legacy/World/ShaderUniforms.cs
@@ -23,7 +23,8 @@ public struct ShaderUniforms(
     RenderLightMode lightMode,
     float gammaCorrection,
     int maxDistance,
-    bool useBrightmaps)
+    bool useBrightmaps,
+    int downScaleAmount)
 {
     public mat4 Mvp = mvp;
     public mat4 MvpNoPitch = mvpNoPitch;
@@ -40,4 +41,5 @@ public struct ShaderUniforms(
     public float GammaCorrection = gammaCorrection;
     public int MaxDistance = maxDistance;
     public bool UseBrightmaps = useBrightmaps;
+    public int DownScaleAmount = downScaleAmount;
 }